+++ src/loss.c
@@ -8,7 +8,7 @@
 float mse_loss(const float *y, const float *t, const size_t size) {
     float loss = 0.0f;
 
-    for (int i = 0; i < size; i++) {
+    for (size_t i = 0; i < size; i++) {
         loss += (t[i] - y[i]) * (t[i] - y[i]);
     }
 
@@ -20,7 +20,7 @@ float mse_loss(const float *y, const float *t, const size_t size) {
 float se_loss(const float *y, const float *t, const size_t size) {
     float loss = 0.0f;
 
-    for (int i = 0; i < size; i++) {
+    for (size_t i = 0; i < size; i++) {
         loss += (t[i] - y[i]) * (t[i] - y[i]);
     }
